% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange-humanorder.R
\name{arrange_humanorder}
\alias{arrange_humanorder}
\title{Sort a data frame by an ID column like F1, F11, F100
so that F11 comes before F100
Sort a data frame by an ID column like F1, F11, F100
using gtools::mixedorder(), which is safer than mixedsort() for data frames}
\usage{
arrange_humanorder(df, col)
}
\description{
Sort a data frame by an ID column like F1, F11, F100
so that F11 comes before F100
Sort a data frame by an ID column like F1, F11, F100
using gtools::mixedorder(), which is safer than mixedsort() for data frames
}
