% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-coat-color.R
\name{call_diplotypes}
\alias{call_diplotypes}
\title{Call the most probable marginal diplotype at a single marker.}
\usage{
call_diplotypes(pr)
}
\arguments{
\item{pr}{A samples-by-founders matrix of allele probabilities at one marker.}
}
\value{
data.frame with sample ID, two-letter diplotype, and founder probs.
}
\description{
Originally written by Daniel Gatti (dan.gatti@jax.org).
}
\keyword{internal}
