% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoprobs-compute-similarity.R
\name{get_gp_samples}
\alias{get_gp_samples}
\title{Extract sample names from a calc_genoprob object}
\usage{
get_gp_samples(gp, chr = NULL)
}
\description{
Extract sample names from a calc_genoprob object
}
\keyword{internal}
