% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{resolve_col_markers}
\alias{resolve_col_markers}
\title{Resolve and normalize marker/map key columns to canonical names.
Returns a data.frame with marker_id, chr, and pos column names.}
\usage{
resolve_col_markers(df)
}
\description{
Resolve and normalize marker/map key columns to canonical names.
Returns a data.frame with marker_id, chr, and pos column names.
}
