% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{resolve_col_samples}
\alias{resolve_col_samples}
\title{Resolve and normalize sample key column to canonical sample_id.
Returns a data.frame with the renamed column.}
\usage{
resolve_col_samples(df)
}
\description{
Resolve and normalize sample key column to canonical sample_id.
Returns a data.frame with the renamed column.
}
