% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{fix_cols}
\alias{fix_cols}
\title{Wrapper around janitor::clean_names to standardize column names.
Returns a standardized data.frame with cleaned column names.}
\usage{
fix_cols(df)
}
\description{
Wrapper around janitor::clean_names to standardize column names.
Returns a standardized data.frame with cleaned column names.
}
